+++ Infrastructure/EntityFramework/ManufacturingDB/RequestOutsourcePart.cs
@@ -7,7 +7,7 @@ namespace VErp.Infrastructure.EF.ManufacturingDB
     {
         public int RequestOutsourcePartId { get; set; }
         public string RequestOutsourcePartCode { get; set; }
-        public int ProductionOrderDetailId { get; set; }
+        public long ProductionOrderDetailId { get; set; }
         public DateTime DateRequiredComplete { get; set; }
         public int ProductId { get; set; }
         public int Quanity { get; set; }
